pyproject.toml
[project]
name = "batikrvgan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
    "gdown>=5.2.0",
    "granian>=1.6.4",
    "onnxruntime>=1.20.1",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "whitenoise>=6.8.2",
]

[project.scripts]
start-dev = "scripts:run_dev" 
